# swagger-spec.json
{"openapi":"3.0.0","paths":{"/logs":{"get":{"operationId":"LogsController_getLogs","parameters":[{"name":"page","required":true,"in":"query","schema":{"type":"number"}},{"name":"limit","required":true,"in":"query","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["logs"],"security":[{"bearer":[]}]}},"/auth/register":{"post":{"operationId":"AuthController_register","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterDataDto"}}}},"responses":{"201":{"description":""}},"tags":["auth"]}},"/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginLocalDataDto"}}}},"responses":{"200":{"description":""}},"tags":["auth"]}},"/auth/login-firebase":{"post":{"operationId":"AuthController_loginWithFirebaseToken","parameters":[{"name":"authorization","required":true,"in":"header","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["auth"]}},"/auth/verify-token":{"post":{"operationId":"AuthController_verifyToken","parameters":[],"responses":{"200":{"description":""}},"tags":["auth"],"security":[{"bearer":[]}]}},"/auth/verify-admin-token":{"post":{"operationId":"AuthController_verifyAdminToken","parameters":[],"responses":{"200":{"description":""}},"tags":["auth"],"security":[{"bearer":[]}]}},"/auth/admin/all-accounts":{"get":{"operationId":"AuthController_getAllAccounts","parameters":[],"responses":{"200":{"description":""}},"tags":["auth"],"security":[{"bearer":[]}]}},"/auth/admin/create-account":{"post":{"operationId":"AuthController_createAccount","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateAccountDto"}}}},"responses":{"201":{"description":""}},"tags":["auth"],"security":[{"bearer":[]}]}},"/auth/admin/update-account/{id}":{"put":{"operationId":"AuthController_updateAccount","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AdminUpdateAccountDataDto"}}}},"responses":{"200":{"description":""}},"tags":["auth"],"security":[{"bearer":[]}]}},"/auth/admin/delete-account/{id}":{"delete":{"operationId":"AuthController_deleteAccount","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["auth"],"security":[{"bearer":[]}]}},"/auth/user/change-password":{"post":{"operationId":"AuthController_changePassword","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PasswordDto"}}}},"responses":{"200":{"description":""}},"tags":["auth"],"security":[{"bearer":[]}]}},"/auth/user/update-phone":{"post":{"operationId":"AuthController_changePhoneNumber","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PhoneDto"}}}},"responses":{"200":{"description":""}},"tags":["auth"],"security":[{"bearer":[]}]}},"/auth/user/update-username":{"post":{"operationId":"AuthController_changeUsername","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UsernameDto"}}}},"responses":{"200":{"description":""}},"tags":["auth"],"security":[{"bearer":[]}]}},"/file/upload":{"post":{"operationId":"FileController_uploadFile","parameters":[],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"file":{"type":"string","format":"binary"}}}}}},"responses":{"201":{"description":""}},"tags":["file"],"security":[{"bearer":[]}]}},"/file/image/{path}":{"get":{"operationId":"FileController_showFile","parameters":[{"name":"path","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["file"]}},"/project":{"get":{"operationId":"ProjectController_getAll","parameters":[],"responses":{"200":{"description":""}},"tags":["project"],"security":[{"bearer":[]}]},"post":{"operationId":"ProjectController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateProjectDto"}}}},"responses":{"201":{"description":""}},"tags":["project"],"security":[{"bearer":[]}]}},"/project/{id}":{"get":{"operationId":"ProjectController_getOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["project"],"security":[{"bearer":[]}]},"put":{"operationId":"ProjectController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateProjectDto"}}}},"responses":{"200":{"description":""}},"tags":["project"],"security":[{"bearer":[]}]},"delete":{"operationId":"ProjectController_delete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["project"],"security":[{"bearer":[]}]}},"/project/undelete/{id}":{"put":{"operationId":"ProjectController_unDelete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["project"],"security":[{"bearer":[]}]}},"/ticket":{"get":{"operationId":"TicketController_getAll","parameters":[],"responses":{"200":{"description":""}},"tags":["ticket"],"security":[{"bearer":[]}]},"post":{"operationId":"TicketController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTicketDto"}}}},"responses":{"201":{"description":""}},"tags":["ticket"],"security":[{"bearer":[]}]}},"/ticket/{id}":{"get":{"operationId":"TicketController_getOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["ticket"],"security":[{"bearer":[]}]},"put":{"operationId":"TicketController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateTicketDto"}}}},"responses":{"200":{"description":""}},"tags":["ticket"],"security":[{"bearer":[]}]},"delete":{"operationId":"TicketController_delete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["ticket"],"security":[{"bearer":[]}]}},"/ticket/increase/{id}":{"patch":{"operationId":"TicketController_increase","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateTicketQuantityDto"}}}},"responses":{"200":{"description":""}},"tags":["ticket"],"security":[{"bearer":[]}]}},"/ticket/decrease/{id}":{"patch":{"operationId":"TicketController_decrease","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateTicketQuantityDto"}}}},"responses":{"200":{"description":""}},"tags":["ticket"],"security":[{"bearer":[]}]}},"/ticket/undelete/{id}":{"put":{"operationId":"TicketController_unDelete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["ticket"],"security":[{"bearer":[]}]}},"/ticket-user":{"get":{"operationId":"TicketUserController_getAll","parameters":[],"responses":{"200":{"description":""}},"tags":["ticket-user"],"security":[{"bearer":[]}]}},"/ticket-user/{id}":{"get":{"operationId":"TicketUserController_getOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["ticket-user"],"security":[{"bearer":[]}]}},"/ticket-voucher":{"get":{"operationId":"TicketVoucherController_getAll","parameters":[],"responses":{"200":{"description":""}},"tags":["ticket-voucher"],"security":[{"bearer":[]}]},"post":{"operationId":"TicketVoucherController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTicketVoucherDto"}}}},"responses":{"201":{"description":""}},"tags":["ticket-voucher"],"security":[{"bearer":[]}]}},"/ticket-voucher/{id}":{"get":{"operationId":"TicketVoucherController_getOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["ticket-voucher"],"security":[{"bearer":[]}]},"put":{"operationId":"TicketVoucherController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateTicketVoucherDto"}}}},"responses":{"200":{"description":""}},"tags":["ticket-voucher"],"security":[{"bearer":[]}]},"delete":{"operationId":"TicketVoucherController_delete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["ticket-voucher"],"security":[{"bearer":[]}]}},"/ticket-voucher/undelete/{id}":{"put":{"operationId":"TicketVoucherController_unDelete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["ticket-voucher"],"security":[{"bearer":[]}]}},"/ticket-voucher-user/get-available-vouchers":{"post":{"operationId":"TicketVoucherUserController_getAvailableVouchers","parameters":[],"requestBody":{"required":false,"content":{"application/json":{"schema":{"type":"object","properties":{"ticketIds":{"type":"array","items":{"type":"string"}}}}}}},"responses":{"201":{"description":""}},"tags":["ticket-voucher-user"],"security":[{"bearer":[]}]}},"/ticket-order":{"get":{"operationId":"TicketOrderController_getAll","parameters":[],"responses":{"200":{"description":""}},"tags":["ticket-order"],"security":[{"bearer":[]}]},"post":{"operationId":"TicketOrderController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTicketOrderDto"}}}},"responses":{"201":{"description":""}},"tags":["ticket-order"],"security":[{"bearer":[]}]}},"/ticket-order/{id}":{"get":{"operationId":"TicketOrderController_getOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["ticket-order"],"security":[{"bearer":[]}]},"delete":{"operationId":"TicketOrderController_delete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["ticket-order"],"security":[{"bearer":[]}]}},"/ticket-order/undelete/{id}":{"put":{"operationId":"TicketOrderController_unDelete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["ticket-order"],"security":[{"bearer":[]}]}},"/ticket-order-user/create-link":{"post":{"operationId":"TicketOrderUserController_createLinkPayment","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTicketOrderDto"}}}},"responses":{"201":{"description":""}},"tags":["ticket-order-user"]}},"/ticket-order-user/get-result/{orderId}":{"get":{"operationId":"TicketOrderUserController_getResult","parameters":[{"name":"orderId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["ticket-order-user"]}},"/ticket-order-user/get-orders":{"get":{"operationId":"TicketOrderUserController_getResultByEmail","parameters":[],"responses":{"200":{"description":""}},"tags":["ticket-order-user"],"security":[{"bearer":[]}]}},"/ticket-order-checkin/add-checkin-record":{"post":{"operationId":"TicketOrderCheckinController_addCheckinRecord","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AddCheckinRecordDto"}}}},"responses":{"201":{"description":""}},"tags":["ticket-order-checkin"],"security":[{"bearer":[]}]}},"/ticket-order-checkin/get-checkin-records":{"post":{"operationId":"TicketOrderCheckinController_getCheckinRecords","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetCheckinRecordsDto"}}}},"responses":{"201":{"description":""}},"tags":["ticket-order-checkin"],"security":[{"bearer":[]}]}},"/ticket-order-checkin/get-checkin-record-by-id/{id}":{"post":{"operationId":"TicketOrderCheckinController_getCheckinRecordById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":""}},"tags":["ticket-order-checkin"],"security":[{"bearer":[]}]}}},"info":{"title":"Skillcetera - API - Documentation","description":"API for user interface development","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"RegisterDataDto":{"type":"object","properties":{"username":{"type":"string"},"email":{"type":"string"},"phone":{"type":"string"},"password":{"type":"string"}},"required":["username","email","phone","password"]},"LoginLocalDataDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}},"required":["email","password"]},"CreateAccountDto":{"type":"object","properties":{"username":{"type":"string"},"email":{"type":"string"},"phone":{"type":"string"},"password":{"type":"string"},"role":{"type":"string"}},"required":["username","email","phone","password","role"]},"AdminUpdateAccountDataDto":{"type":"object","properties":{"username":{"type":"string"},"phone":{"type":"string"},"role":{"type":"string"},"password":{"type":"string"}},"required":["username","phone","role","password"]},"PasswordDto":{"type":"object","properties":{"value":{"type":"string"}},"required":["value"]},"PhoneDto":{"type":"object","properties":{"value":{"type":"string"}},"required":["value"]},"UsernameDto":{"type":"object","properties":{"value":{"type":"string"}},"required":["value"]},"CreateProjectDto":{"type":"object","properties":{"projectName":{"type":"string"},"description":{"type":"string"},"startDate":{"format":"date-time","type":"string"},"endDate":{"format":"date-time","type":"string"}},"required":["projectName","description","startDate","endDate"]},"UpdateProjectDto":{"type":"object","properties":{"projectName":{"type":"string"},"description":{"type":"string"},"startDate":{"format":"date-time","type":"string"},"endDate":{"format":"date-time","type":"string"}},"required":["projectName","description","startDate","endDate"]},"CreateTicketDto":{"type":"object","properties":{"ticketName":{"type":"string"},"description":{"type":"string"},"price":{"type":"number"},"quantity":{"type":"number"},"startDate":{"format":"date-time","type":"string"},"endDate":{"format":"date-time","type":"string"},"project":{"type":"string"}},"required":["ticketName","description","price","quantity","startDate","endDate","project"]},"UpdateTicketDto":{"type":"object","properties":{"ticketName":{"type":"string"},"description":{"type":"string"},"price":{"type":"number"},"startDate":{"format":"date-time","type":"string"},"endDate":{"format":"date-time","type":"string"}},"required":["ticketName","description","price","startDate","endDate"]},"UpdateTicketQuantityDto":{"type":"object","properties":{"quantity":{"type":"number"}},"required":["quantity"]},"CreateTicketVoucherDto":{"type":"object","properties":{"voucherCode":{"type":"string"},"description":{"type":"string"},"discount":{"type":"number"},"quantity":{"type":"number"},"startDate":{"format":"date-time","type":"string"},"endDate":{"format":"date-time","type":"string"},"applyTicketId":{"type":"array","items":{"type":"string"}},"note":{"type":"string"}},"required":["voucherCode","description","discount","quantity","startDate","endDate","applyTicketId","note"]},"UpdateTicketVoucherDto":{"type":"object","properties":{"ticketName":{"type":"string"},"description":{"type":"string"},"price":{"type":"number"},"quantity":{"type":"number"},"startDate":{"format":"date-time","type":"string"},"endDate":{"format":"date-time","type":"string"},"note":{"type":"string"}},"required":["ticketName","description","price","quantity","startDate","endDate","note"]},"CreateTicketOrderDto":{"type":"object","properties":{"email":{"type":"string"},"phone":{"type":"string"},"username":{"type":"string"},"total":{"type":"number"},"ticketVoucher":{"type":"object","description":"Ticket Voucher ID","default":""},"items":{"type":"array","items":{"type":"object","properties":{"ticketId":{"type":"string"},"name":{"type":"string"},"quantity":{"type":"number"},"price":{"type":"number"}}}}},"required":["email","phone","username","total","items"]},"AddCheckinRecordDto":{"type":"object","properties":{"idOrder":{"type":"string"},"idItem":{"type":"string"}},"required":["idOrder","idItem"]},"GetCheckinRecordsDto":{"type":"object","properties":{"idOrder":{"type":"string"}},"required":["idOrder"]}}}}